Refactor code: re-arrange directory structure for store and api endpoints
